+++ Platon/backend/tests/profile_management_test.py
@@ -3,7 +3,8 @@ from tests.base_test import TestConfig
 from app.auth_system.models import User
 from app.profile_management.models import ResearchInformation
 from app.auth_system.views import generate_token
-from app.profile_management.views import ResearchType,ResearchInfoFetch
+from app.profile_management.views import ResearchType
+from app.profile_management.helpers import ResearchInfoFetch
 from app.follow_system.models import Follow
 from app import db
 import jwt
